# services/pizza-service/docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: OCI Pizza - Pizza Service API
  description: |
    API responsável pelo gerenciamento do catálogo de pizzas da aplicação OCI Pizza.

    O serviço permite consultar e gerenciar informações relacionadas ao catálogo
    de pizzas.

    As operações de alteração do catálogo requerem autenticação administrativa.
  version: 1.0.0

servers:
  - url: http://localhost:8001
    description: Ambiente local

  - url: https://api-dev.ocipizza.com/pizza-service
    description: Ambiente de desenvolvimento

  - url: https://api-hml.ocipizza.com/pizza-service
    description: Ambiente de homologação

  - url: https://api.ocipizza.com/pizza-service
    description: Ambiente de produção

tags:
  - name: Pizzas
    description: Operações relacionadas ao catálogo de pizzas
paths:
  /pizzas:
    get:
      tags:
        - Pizzas
      summary: Lista pizzas do catálogo
      description: |
        Retorna pizzas disponíveis no catálogo.
        Esta operação pode ser utilizada por consumidores autorizados
        da aplicação.
      responses:
        "200":
          description: Lista de pizzas retornada com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PizzaListResponse"

        "500":
          description: Erro interno do serviço.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

    post:
      tags:
        - Pizzas
      summary: Cadastra uma nova pizza
      description: |
        Cria uma nova pizza no catálogo.
        Esta operação requer permissão administrativa.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PizzaCreateRequest"
      responses:
        "201":
          description: Pizza criada com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PizzaResponse"

        "400":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendFail"

        "401":
          description: Usuário não autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

        "403":
          description: Usuário sem permissão administrativa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

  /pizzas/{id}:
    get:
      tags:
        - Pizzas
      summary: Consulta uma pizza específica
      description: |
        Retorna uma pizza através do identificador.
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único da pizza
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        "200":
          description: Pizza encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PizzaResponse"

        "404":
          description: Pizza não encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendFail"

    patch:
      tags:
        - Pizzas
      summary: Atualiza informações de uma pizza
      description: |
        Atualiza informações existentes de uma pizza.
        Esta operação requer permissão administrativa.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único da pizza
          schema:
            type: integer
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PizzaUpdateRequest"
      responses:
        "200":
          description: Pizza atualizada com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PizzaResponse"

        "400":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendFail"

        "401":
          description: Usuário não autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

        "403":
          description: Usuário sem permissão administrativa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

        "404":
          description: Pizza não encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendFail"

    delete:
      tags:
        - Pizzas
      summary: Remove uma pizza do catálogo
      description: |
        Remove uma pizza do catálogo.
        Esta operação requer permissão administrativa.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único da pizza
          schema:
            type: integer
            minimum: 1
            example: 1

      responses:
        "200":
          description: Pizza removida com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendSuccess"

        "401":
          description: Usuário não autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

        "403":
          description: Usuário sem permissão administrativa.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendError"

        "404":
          description: Pizza não encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSendFail"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Pizza:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
          example: 1
        name:
          type: string
          example: Margherita
        description:
          type: string
          example: Molho de tomate, queijo e manjericão
        category:
          type: string
          example: tradicional
        price:
          type: number
          format: float
          example: 39.90
        image_name:
          type: string
          example: margherita.jpg
        available:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PizzaCreateRequest:
      type: object
      required:
        - name
        - category
        - price
      properties:
        name:
          type: string
          example: Margherita
        description:
          type: string
          example: Molho de tomate, queijo e manjericão
        category:
          type: string
          example: tradicional
        price:
          type: number
          format: float
          example: 39.90
        image_name:
          type: string
          example: margherita.jpg
        available:
          type: boolean
          default: true

    PizzaUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        image_name:
          type: string
        available:
          type: boolean

    PizzaResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            pizza:
              $ref: "#/components/schemas/Pizza"

    PizzaListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            pizzas:
              type: array
              items:
                $ref: "#/components/schemas/Pizza"

    JSendSuccess:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          nullable: true

    JSendFail:
      type: object
      properties:
        status:
          type: string
          example: fail
        data:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_REQUIRED
            field:
              type: string
              example: name
            message:
              type: string
              example: Field is required

    JSendError:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Internal server error